// src/pages/dashboard/types.ts
export interface IGridPos {
  h: number;
  w: number;
  x: number;
  y: number;
  i: string;
}

// query interface
export interface ITarget {
  refId: string;
  __mode__: '__expr__' | '__query__';
  expr: string; // promQL
  legendFormat: string;
  time?: IRawTimeRange; // 固定时间范围，2025-10-20 废弃
  step?: number; // 2024-01-24 从固定 step 改成 min step (v7)
  maxDataPoints?: number; // 2024-01-24 新增 maxDataPoints 用于计算默认的 step (v7)，2025-10-20 废弃
  query?: {
    index: string;
    index_type: 'index' | 'index_pattern';
    filters: string;
    values: {
      func: string;
    }[];
  };
  legend?: string;
  instant?: boolean;
  hide?: boolean;
}

export type IType = 'row' | 'timeseries' | 'stat' | 'table' | 'tableNG' | 'pie' | 'hexbin' | 'barGauge' | 'text' | 'gauge' | 'iframe';

export interface IValueMapping {
  match: {
    special?: string | number;
    specialValue?: string | number;
    from?: number;
    to?: number;
  };
  result: {
    color: string;
    text: string;
  };
  type: 'range' | 'special' | 'specialValue'; // TODO: 历史原因 special 是固定值，specialValue 是特殊值
}

export interface IThresholds {
  steps: {
    color: string;
    value: number;
    type?: 'base';
  }[];
  mode: 'absolute' | 'percentage';
}

export interface ThresholdsStyle {
  mode: 'off' | 'line' | 'dashed' | 'line+area' | 'dashed+area';
}

export interface LinksItem {
  title: string;
  url: string;
  targetBlank: boolean;
}

export interface IStandardOptions {
  unit?: string; // 纠正的单位
  min?: number;
  max?: number;
  decimals?: number;
  dateFormat?: string;
  displayName?: string;
}

// 一些通用的配置，不同类型的图表可选择性使用配置
export interface IOptions {
  valueMappings?: IValueMapping[];
  thresholds?: IThresholds;
  thresholdsStyle?: ThresholdsStyle;
  xThresholds?: IThresholds;
  standardOptions?: IStandardOptions;
  legend?: {
    // TODO: 目前不支持这么复杂的自定义
    calcs: string[];
    displayMode: 'list' | 'table' | 'hidden';
    placement: 'right' | 'bottom';
    heightInPercentage?: number;
    widthInPercentage?: number;
    columns?: string[];
    detailName: string;
    detailUrl: string;
    behaviour: 'showItem' | 'hideItem';
    selectMode: 'single' | 'multiple';
  };
  tooltip?: {
    mode: 'single' | 'all';
    sort: 'none' | 'asc' | 'desc';
  };
  colors?: {
    scheme: string;
  };
  links?: LinksItem[];
}

export interface IOverride {
  matcher: {
    type: 'byName'; // 目前只支持 byName
    value: string;
  };
  properties: {
    [key: string]: any; // standardOptions | valueMappings
  };
}

export interface ILink {
  type: 'link' | 'dashboards';
  title: string;
  url: string;
  targetBlank?: boolean;
  dashboardIds?: number[];
  dashboards: {
    id: number;
    name: string;
    ident: string;
  }[];
}
// ...
export interface ITransformation {
  id: 'string';
  options: {
    [key: string]: any;
  };
  disabled?: boolean;
}

export interface IPanel {
  version: string; // 单个图表面板使用的版本
  id: string;
  name: string;
  links?: ILink[];
  description: string;
  layout: IGridPos;
  datasourceCate?: string; // 5.11.0 新增支持配置数据源类型，默认是 prometheus
  datasourceValue?: number; // 6.x 开始 datasourceName 已经废弃，datasourceValue 即 datasourceId
  targets: ITarget[];
  type: IType;
  options: IOptions;
  custom: any; // 图表
  overrides: IOverride[];
  collapsed?: boolean; // 用于 row 展开收起控制是否显示
  panels?: IPanel[]; // 用于 row 收起时保存子面板
  transformations?: ITransformation[];
  transformationsNG?: ITransformation[]; // NG 版本的变换
  repeat?: string;
  maxPerRow?: number;
  repeatPanelId?: string;
  scopedVars?: any;
  maxDataPoints?: number; // 2025-10-20 新增
  queryOptionsTime?: IRawTimeRange; // 2025-10-20 新增， queryOptionsTime 会覆盖 time
}